// crypto-vault-expo/src/core/services/TonServices/type.ts
export interface Account {
    /**
     * @format address
     * @example "0:da6b1b6663a0e4d18cc8574ccd9db5296e367dd9324706f3bbd9eb1cd2caf0bf"
     */
    address: Address;
    /**
     * @format bigint
     * @example 123456789
     */
    balance: bigint;
    /**
     * {'USD': 1, 'IDR': 1000}
     * @example {}
     */
    currencies_balance?: Record<string, any>;
    /**
     * unix timestamp
     * @format int64
     * @example 1720860269
     */
    last_activity: number;
    status: AccountStatus;
    interfaces?: string[];
    /** @example "Ton foundation" */
    name?: string;
    /** @example true */
    is_scam?: boolean;
    /** @example "https://ton.org/logo.png" */
    icon?: string;
    /** @example true */
    memo_required?: boolean;
    /** @example ["get_item_data"] */
    get_methods: string[];
    is_suspended?: boolean;
    is_wallet: boolean;
}

export interface TonAccountsType extends Account {}